Add an explicit cast to output values instead of GlobRefs in examples. (one more that got left out)

+++ dash/examples/ex.02.array-local/main.cpp
@@ -21,7 +21,7 @@ int main(int argc, char* argv[])
   arr.barrier();
   if (static_cast<size_t>(myid) == size-1) {
     for (auto el: arr) {
-      cout << el << " ";
+      cout << (int)el << " ";
     }
     cout << endl;
   }
